fix: resolve all pre-existing mypy type errors across codebase
- Add type annotations to all function signatures in utils/ and tests/
- Wrap numpy np.mean() returns with float() for strict typing
- Fix variable shadowing (result→test_result, fmt→console_fmt, metric→unique names)
- Use Optional[Callable] and explicit typed intermediates for no-any-return
- Use importlib.util.find_spec() for availability checks (avoids F401)
- Remove unused Dict import from ollama_provider

+++ src/scoring/pass_k.py
@@ -107,7 +107,7 @@ def calculate_pass_at_k_batch(results: List[List[bool]], k_values: List[int]) ->
                 score = calculate_pass_at_k(n, c, k)
                 scores[k].append(score)
 
-    return {k: np.mean(v) if v else 0.0 for k, v in scores.items()}
+    return {k: float(np.mean(v)) if v else 0.0 for k, v in scores.items()}
 
 
 def evaluate_code_samples(
